requirements.txt
numpy==1.19.5
pandas==1.2.4
streamlit
openpyxl
pyyaml
matplotlib